web-extension: use proper number format for position dom properties
DOM bindings use double, not long. This fixes the form filler
popup not showing up. Use the proper DOM bindings API instead
of g_object_get() to avoid this happening again.

https://bugzilla.gnome.org/show_bug.cgi?id=750842

+++ lib/ephy-web-dom-utils.h
@@ -44,12 +44,12 @@ gboolean ephy_web_dom_utils_find_form_auth_elements (WebKitDOMHTMLFormElement *f
                                                      WebKitDOMNode           **password);
 
 void ephy_web_dom_utils_get_absolute_bottom_for_element (WebKitDOMElement *element,
-                                                         long             *x,
-                                                         long             *y);
+                                                         double           *x,
+                                                         double           *y);
 
 void ephy_web_dom_utils_get_absolute_position_for_element(WebKitDOMElement *element,
-                                                          long             *x,
-                                                          long             *y);
+                                                          double           *x,
+                                                          double           *y);
 
 char *ephy_web_dom_utils_get_selection_as_string (WebKitDOMDOMSelection *selection);
 G_END_DECLS
